chore: move common code to helper.go
